split up files with multiple utils
