+++ shared-data/js/helpers/getStackerLocationFromSlotName.ts
@@ -1,0 +1,3 @@
+export function getStackerLocationFromSlotName(slotName: string): string {
+  return `STACKER ${slotName.charAt(0)}`
+}
