+++ shared-data/js/helpers/getTopLabwareFromStack.ts
@@ -1,0 +1,10 @@
+import type { LabwareInStack, StackItem } from '../types'
+
+export function getTopLabwareFromStack(
+  itemsOnDeck: StackItem[]
+): LabwareInStack | null {
+  const topLabwareInStack = itemsOnDeck.find(
+    (stackedItem): stackedItem is LabwareInStack => 'labwareId' in stackedItem
+  )
+  return topLabwareInStack ?? null
+}
